Add test to verify docker images build.
Currently failes on a number of the test cases, I'm not currently sure how to programatically fix this however, as it seems to be based on the underlying frameworks requiring certain files or folders to exist in order to complete the application build step.
It might be worth considering pulling the latest version of whatever "starter project" these test candidates offer, removing any Dockerfile possibly shipped with the starter project, building it with the tool then verifying its correctness.